decode: do not count n_hdr in check for minimum data length
The check was wrong. A caller could supply n_in=0 and n_hdr=4, and this
was satisfying the check that the input will decode to at least 4 bytes,
the size of the hash fragment, despite that there is actually no input.
This commit fixes the logic so that the check occurs before n_hdr gets
added to the needed output buffer size.