decode: reduce temp allocation when called with oversized output buffer
We allocate a temporary scratch buffer to receive the bignum limbs from
the base conversion routine, and we were sizing it large enough to hold
data that completely fill the output buffer. That will be minimal in the
case that we dynamically allocated the output buffer ourselves, but the
caller may have supplied their own output buffer that may be much larger
than is necessary to hold the actual decoded data. Instead of sizing our
scratch buffer based on the actual output buffer size, size it based on
the _needed_ output buffer size.