decode: do not invoke Undefined Behavior when handed a zero input
If the caller supplies any input string consisting solely of the digit
'1' repeated at least 4 times (so as to satisfy the check for minimum
output length), then the input to the base conversion routine will be
numerically equal to zero. A minimally sized bignum zero has no limbs at
all, which means that we were calling malloc(0), which isn't a problem
per se, but then we were clearing the "last" limb to zero. There is no
last limb when there are no limbs at all, so we were in fact clobbering
the heap. Worse, we next scan forward through the output of the base
conversion routine to find the first non-zero output byte, but in the
case of a zero input, no non-zero output bytes will have been written,
so we were scanning right off the end of the buffer, and upon finding a
non-zero byte somewhere in memory, we were then doing a memmove with a
source address at that arbitrary byte and with a length of almost
SIZE_MAX, effectively corrupting the process's entire address space (or
until triggering a segfault). The fix is simple: abort early, without
even allocating the scratch buffer, if the number of limbs is zero. We
are assuming that the caller will never supply an input consisting of a
run of exactly enough '1' digits such that the first four bytes of the
double-SHA256 hash are all zeros. There is a probably some number N such
that SHA256(SHA256([0x00**N]))[0..3]==0x00000000, but we're just going
to say that that particular byte string is unlikely enough that we can
safely reject its Base58Check encoding even though it's technically
valid.